kill listing row
