--- client/styles/components/listing_row.scss
@@ -1,79 +1,0 @@
-@import 'client/styles/shared';
-
-.ListingRow {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 100%;
-  cursor: pointer;
-  padding: 16px;
-
-  &:hover {
-    background: $background-color-white-hover;
-  }
-
-  .pinned {
-    align-items: center;
-    border-radius: 12px;
-    border: 1px solid #ddd;
-    display: flex;
-    height: 48px;
-    justify-content: center;
-    margin-right: 24px;
-    width: 48px;
-
-    svg {
-      fill: currentColor;
-    }
-  }
-
-  .reaction {
-    margin-right: 24px;
-  }
-
-  .title-container {
-    align-self: center;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-
-    .row-header {
-      a {
-        color: $text-color-black;
-        font-size: 19px;
-        text-decoration: none;
-        font-weight: 600;
-      }
-    }
-
-    .row-subheader {
-      color: $text-color-light;
-      display: flex;
-
-      a {
-        color: $text-color-light;
-        font-size: 17px;
-        overflow: hidden;
-        text-overflow: ellipsis;
-
-        &:hover {
-          color: darken($text-color-light, 8%);
-        }
-      }
-    }
-  }
-
-  .content-right-container {
-    display: flex;
-    flex-direction: row;
-  }
-
-  @include xs-max {
-    .content-right-container,
-    .activity-icons,
-    .last-active {
-      display: none;
-    }
-  }
-}
